+++ src/config/settings/settings.mjs
@@ -442,6 +442,14 @@ const SETTINGS = {
             scope: "world",
             default: true,
             config: false,
+            requiresReload: true,
+          },
+          "allow-arcane-ward-enhancer": {
+            type: Boolean,
+            scope: "world",
+            default: true,
+            config: false,
+            requiresReload: true,
           },
         },
         IMPORT: {