+++ src/effects/_module.mjs
@@ -1,3 +1,4 @@
 export * as External from "./external/_module.mjs";
 export { default as ExternalAutomations } from "./external/ExternalAutomations.mjs";
 export { default as DDBEffectHelper } from "./DDBEffectHelper.mjs";
+export { DDBEnhancers } from "./enhancers/_module.mjs";